# docs/swagger.yaml
basePath: /1sat
definitions:
...
  merkle.ArcCallback:
    properties:
      merklePath:
        items:
          type: integer
        type: array
      txId:
        type: string
      txStatus:
        type: string
    type: object
...
  pkg_ordfs.Response:
    properties:
      content:
        items:
          type: integer
        type: array
      contentLength:
        type: integer
      contentType:
        type: string
      map:
        items:
          type: integer
        type: array
      origin:
        $ref: '#/definitions/transaction.Outpoint'
      outpoint:
        $ref: '#/definitions/transaction.Outpoint'
      parent:
        $ref: '#/definitions/transaction.Outpoint'
      sequence:
        type: integer
    type: object
...
  transaction.Outpoint:
    properties:
      index:
        type: integer
      txid:
        items:
          type: integer
        type: array
    type: object
info:
  contact:
    name: API Support
    url: https://github.com/b-open-io/1sat-stack
  description: Composable BSV blockchain services API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 1Sat Stack API
  version: "1.0"
paths:
...
  /arc/callback:
    post:
      consumes:
      - application/json
      description: Receives transaction status updates from Arc broadcaster
      parameters:
      - description: Arc callback payload
        in: body
        name: callback
        required: true
        schema:
          $ref: '#/definitions/merkle.ArcCallback'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Handle Arc callback
      tags:
      - merkle
...
  /content/{path}:
    get:
      description: Serve the content of an inscription by outpoint or txid, with directory
        and SPA support
      parameters:
      - description: Outpoint (txid_vout) or txid, optionally with :seq and /filepath
        in: path
        name: path
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Content
          schema:
            type: file
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get inscription content
      tags:
      - ordfs
...
  /ordfs/metadata/{path}:
    get:
      description: Get metadata about inscription content without downloading the
        content
      parameters:
      - description: Outpoint (txid_vout) or txid
        in: path
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Metadata
          schema:
            $ref: '#/definitions/pkg_ordfs.Response'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get content metadata
      tags:
      - ordfs
  /ordfs/preview:
    post:
      consumes:
      - '*/*'
      description: Echo back the request body for preview rendering
      produces:
      - '*/*'
      responses:
        "200":
          description: Content
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Preview posted content
      tags:
      - ordfs
  /ordfs/preview/{b64HtmlData}:
    get:
      description: Decode and render base64-encoded HTML
      parameters:
      - description: Base64-encoded HTML content
        in: path
        name: b64HtmlData
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML content
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Preview HTML content
      tags:
      - ordfs
  /ordfs/stream/{outpoint}:
    get:
      description: Stream content from an ordinal chain
      parameters:
      - description: Outpoint (txid_vout)
        in: path
        name: outpoint
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Streamed content
          schema:
            type: file
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Stream content
      tags:
      - ordfs
...
swagger: "2.0"